requirements.txt
beautifulsoup4
langchain_openai
loguru
neo4j
numpy
pandas
plotly
pydeck
python-dotenv
Requests
scikit_learn
streamlit
torch==2.5.0
torch-geometric
kagglehub
datatable
ogb
torch-scatter 
torch-sparse
pyg-lib
matplotlib